// src/modules/authentication/models/session_user_model.tsx
export interface SessionUserModel {
  id: string;
  idAssinatura: string;
  nome: string;
  email: string;
  root: boolean;
  adm: boolean;
  permissoes: PermissoesUserModel;
}

export interface PermissoesUserModel {
  clientesRead: boolean;
  clientesInsert: boolean;
  clientesUpdate: boolean;
  clientesDelete: boolean;
  produtosRead: boolean;
  produtosInsert: boolean;
  produtosUpdate: boolean;
  produtosDelete: boolean;
  produtosInsertBaixaEstoque: boolean;
  agendaRead: boolean;
  agendaInsert: boolean;
  agendaUpdate: boolean;
  agendaDelete: boolean;
  usuariosRead: boolean;
  usuariosInsert: boolean;
  usuariosUpdate: boolean;
  usuariosDelete: boolean;
  fornecedoresRead: boolean;
  fornecedoresInsert: boolean;
  fornecedoresUpdate: boolean;
  fornecedoresDelete: boolean;
  categoriasProdutosRead: boolean;
  categoriasProdutosInsert: boolean;
  categoriasProdutosUpdate: boolean;
  categoriasProdutosDelete: boolean;
  despesasRead: boolean;
  despesasInsert: boolean;
  despesasUpdate: boolean;
  despesasDelete: boolean;
  contasPagarRead: boolean;
  contasPagarInsert: boolean;
  contasPagarUpdate: boolean;
  contasPagarDelete: boolean;
  contasPagarWriteOff: boolean;
  contasPagarParcelling: boolean;
  contasReceberRead: boolean;
  contasReceberInsert: boolean;
  contasReceberUpdate: boolean;
  contasReceberDelete: boolean;
  contasReceberWriteOff: boolean;
  contasReceberParcelling: boolean;
  vendasRead: boolean;
  vendasInsert: boolean;
  vendasUpdate: boolean;
  vendasDelete: boolean;
  vendasFinalize: boolean;
  hfEscolasRead: boolean;
  hfEscolasInsert: boolean;
  hfEscolasUpdate: boolean;
  hfEscolasDelete: boolean;
  hfContratosRead: boolean;
  hfContratosInsert: boolean;
  hfContratosUpdate: boolean;
  hfContratosDelete: boolean;
  hfContratosWriteOff: boolean;
  marcasProdutosRead: boolean;
  marcasProdutosInsert: boolean;
  marcasProdutosUpdate: boolean;
  marcasProdutosDelete: boolean;
}